move selector package to common
