+++ sensor/kubernetes/listener/resources/deployment_store_bench_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/stackrox/rox/generated/storage"
-	"github.com/stackrox/rox/sensor/kubernetes/selector"
+	"github.com/stackrox/rox/sensor/common/selector"
 )
 
 var (
