+++ sensor/kubernetes/listener/resources/deployment_store_test.go
@@ -8,10 +8,10 @@ import (
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/uuid"
 	"github.com/stackrox/rox/sensor/common/registry"
+	"github.com/stackrox/rox/sensor/common/selector"
 	"github.com/stackrox/rox/sensor/common/service"
 	"github.com/stackrox/rox/sensor/common/store"
 	"github.com/stackrox/rox/sensor/kubernetes/orchestratornamespaces"
-	"github.com/stackrox/rox/sensor/kubernetes/selector"
 	"github.com/stretchr/testify/suite"
 	v1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
